Refactor splunkhttp to make it more universal
