+++ instrumentation/net/http/splunkhttp/README.md
@@ -10,10 +10,9 @@ package main
 import (
 	"net/http"
 
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-
 	"github.com/signalfx/splunk-otel-go/distro"
 	"github.com/signalfx/splunk-otel-go/instrumentation/net/http/splunkhttp"
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 )
 
 func main() {
@@ -22,7 +21,8 @@ func main() {
 	var handler http.Handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("Hello"))
 	})
-	handler = splunkhttp.NewHandler(handler, "my-service")
+	handler = splunkhttp.NewHandler(handler)
+	handler = otelhttp.NewHandler(handler, "my-service")
 
 	http.ListenAndServe(":9090", handler)
 }
@@ -34,7 +34,7 @@ func main() {
 
 | Code                                                       | Environment variable                   | Default value  | Purpose                                         |
 | ---------------------------------------------------------- | -------------------------------------- | -------------- | ----------------------------------------------- |
-| `WithTraceResponseHeader`, `TraceResponseHeaderMiddleware` | `SPLUNK_TRACE_RESPONSE_HEADER_ENABLED` | `true`         | Adds `Server-Timing` header to HTTP responses.  |
+| `WithTraceResponseHeader`, `TraceResponseHeaderMiddleware` | `SPLUNK_TRACE_RESPONSE_HEADER_ENABLED` | `true`         | Adds `Server-Timing` header to HTTP responses. [More](#Trace-linkage-between-the-APM-and-RUM-products) |
 
 ## Features
 
@@ -42,4 +42,11 @@ func main() {
 
 `TraceResponseHeaderMiddleware` wraps the passed handler, functioning like middleware.
 It adds trace context in [traceparent form](https://www.w3.org/TR/trace-context/#traceparent-header)
-as [Server-Timing header](https://www.w3.org/TR/server-timing/) to the HTTP response.
+as [Server-Timing header](https://www.w3.org/TR/server-timing/) to the HTTP response:
+
+```HTTP
+Access-Control-Expose-Headers: Server-Timing
+Server-Timing: traceparent;desc="00-<serverTraceId>-<serverSpanId>-01"
+```
+
+This information can be later consumed by the [splunk-otel-js-web](https://github.com/signalfx/splunk-otel-js-web) library.
