+++ instrumentation/net/http/splunkhttp/config_test.go
@@ -32,7 +32,6 @@ func TestConfigs(t *testing.T) {
 		{
 			name: "Default",
 			assert: func(t *testing.T, c *config) {
-				assert.Nil(t, c.OTelOpts, "should not set any additional OTel options")
 				assert.True(t, c.TraceResponseHeaderEnabled, "should enable TraceResponseHeader")
 			},
 		},
