+++ instrumentation/net/http/splunkhttp/handler_test.go
@@ -19,14 +19,11 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-	"go.opentelemetry.io/otel/oteltest"
 )
 
-func TestTraceResponseHeaderMiddleware(t *testing.T) {
+func TestNewHandlerDefault(t *testing.T) {
 	resp := responseForHandler(func(handler http.Handler) http.Handler { // nolint:bodyclose // Body is not used
-		handler = TraceResponseHeaderMiddleware(handler)
-		return otelhttp.NewHandler(handler, "server", otelhttp.WithTracerProvider(oteltest.NewTracerProvider()))
+		return NewHandler(handler)
 	})
 
 	assert.Equal(t, http.StatusOK, resp.StatusCode, "should return OK status code")
@@ -34,12 +31,12 @@ func TestTraceResponseHeaderMiddleware(t *testing.T) {
 	assert.Regexp(t, "^traceparent;desc=\"00-[0-9a-f]{32}-[0-9a-f]{16}-01\"$", resp.Header.Get("Server-Timing"), "should return properly formated Server-Timing header")
 }
 
-func TestNewHandlerDefault(t *testing.T) {
+func TestNewHandlerTraceResponseHeaderDisabled(t *testing.T) {
 	resp := responseForHandler(func(handler http.Handler) http.Handler { // nolint:bodyclose // Body is not used
-		return NewHandler(handler, "server", WithOTelOpts(otelhttp.WithTracerProvider(oteltest.NewTracerProvider())))
+		return NewHandler(handler, WithTraceResponseHeader(false))
 	})
 
 	assert.Equal(t, http.StatusOK, resp.StatusCode, "should return OK status code")
-	assert.Contains(t, resp.Header["Access-Control-Expose-Headers"], "Server-Timing", "should set Access-Control-Expose-Headers header")
-	assert.Regexp(t, "^traceparent;desc=\"00-[0-9a-f]{32}-[0-9a-f]{16}-01\"$", resp.Header.Get("Server-Timing"), "should return properly formated Server-Timing header")
+	assert.NotContains(t, resp.Header["Access-Control-Expose-Headers"], "Server-Timing", "should NOT set Access-Control-Expose-Headers header")
+	assert.NotRegexp(t, "^traceparent;desc=\"00-[0-9a-f]{32}-[0-9a-f]{16}-01\"$", resp.Header.Get("Server-Timing"), "should not add traceID to Server-Timing header")
 }
